chore: rename test_config.py
